+++ descarteslabs/__init__.py
@@ -12,18 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# flake8: noqa
 
-__path__ = __import__('pkgutil').extend_path(__path__, __name__)
-__version__ = "0.5.0"
-
-from .auth import Auth
-descartes_auth = Auth.from_environment_or_token_json()
-
-from .services.metadata import Metadata
-from .services.places import Places
-from .services.raster import Raster
-
-metadata = Metadata()
-places = Places()
-raster = Raster()
+__version__ = "0.5.0"  # noqa
