+++ descarteslabs/services/metadata/metadata.py
@@ -16,11 +16,11 @@ import json
 import os
 from warnings import warn, simplefilter
 from six import string_types
-from .service import Service
-from .places import Places
-import descarteslabs as dl
-
-from . import metadata_filtering as filtering
+from descarteslabs.services.base import Service
+from descarteslabs.services.places import Places
+from descarteslabs.auth import Auth
+from descarteslabs.services.raster import Raster
+from descarteslabs.services.metadata.metadata_filtering import AndExpression
 
 CONST_ID_DEPRECATION_MESSAGE = (
     "Keyword arg `const_id' has been deprecated and will be removed in "
@@ -47,16 +47,18 @@ class Metadata(Service):
 
     TIMEOUT = (9.5, 120)
 
-    def __init__(self, url=None, token=None, auth=dl.descartes_auth):
+    def __init__(self, url=None, token=None, auth=Auth()):
         """The parent Service class implements authentication and exponential
         backoff/retry. Override the url parameter to use a different instance
         of the backing service.
         """
         simplefilter('always', DeprecationWarning)
         if url is None:
-            url = os.environ.get("DESCARTESLABS_METADATA_URL", "https://platform.descarteslabs.com/metadata/v1")
+            url = os.environ.get("DESCARTESLABS_METADATA_URL",
+                                 "https://platform.descarteslabs.com/metadata/v1")
 
         Service.__init__(self, url, token, auth)
+        self._raster = Raster(auth=self.auth)
 
     def sources(self):
         """Get a list of image sources.
@@ -94,7 +96,8 @@ class Metadata(Service):
 
 
         """
-        params = ['limit', 'offset', 'products', 'wavelength', 'resolution', 'tags']
+        params = ['limit', 'offset', 'products',
+                  'wavelength', 'resolution', 'tags']
 
         args = locals()
         kwargs = dict(kwargs, **{
@@ -249,7 +252,7 @@ class Metadata(Service):
 
         if dltile is not None:
             if isinstance(dltile, string_types):
-                dltile = dl.raster.dltile(dltile)
+                dltile = self._raster.dltile(dltile)
             if isinstance(dltile, dict):
                 geom = dltile['geometry']
 
@@ -304,7 +307,7 @@ class Metadata(Service):
         if q is not None:
             if not isinstance(q, list):
                 q = [q]
-            kwargs['query_expr'] = filtering.AndExpression(q).serialize()
+            kwargs['query_expr'] = AndExpression(q).serialize()
 
         if pixels:
             kwargs['pixels'] = pixels
@@ -360,7 +363,7 @@ class Metadata(Service):
 
         if dltile is not None:
             if isinstance(dltile, string_types):
-                dltile = dl.raster.dltile(dltile)
+                dltile = self._raster.dltile(dltile)
             if isinstance(dltile, dict):
                 geom = dltile['geometry']
 
@@ -417,7 +420,7 @@ class Metadata(Service):
         if q is not None:
             if not isinstance(q, list):
                 q = [q]
-            kwargs['query_expr'] = filtering.AndExpression(q).serialize()
+            kwargs['query_expr'] = AndExpression(q).serialize()
 
         if sort_field is not None:
             kwargs['sort_field'] = sort_field
@@ -436,7 +439,8 @@ class Metadata(Service):
         fc = {'type': 'FeatureCollection', "features": r.json()}
 
         if 'x-continuation-token' in r.headers:
-            fc['properties'] = {'continuation_token': r.headers['x-continuation-token']}
+            fc['properties'] = {
+                'continuation_token': r.headers['x-continuation-token']}
 
         return fc
