+++ descarteslabs/services/raster/__init__.py
@@ -1,0 +1,3 @@
+from .raster import Raster
+
+__all__ = ["Raster"]
