+++ descarteslabs/utilities.py
@@ -17,16 +17,6 @@ import struct
 
 from .addons import blosc, numpy as np
 from .exceptions import ServerError
-from .services.metadata_filtering import Properties
-
-properties = Properties("absolute_orbit", "acquired", "archived", "area", "azimuth_angle",
-                        "bright_fraction", "cirrus_fraction", "cloud_fraction", "cloud_fraction_0",
-                        "cloud_fraction_1", "descartes_version", "earth_sun_distance",
-                        "fill_fraction", "geolocation_accuracy", "opaque_fraction", "product", "processed",
-                        "published", "relative_orbit", "roll_angle", "sat_id", "solar_azimuth_angle",
-                        "solar_azimuth_angle_0", "solar_elevation_angle", "solar_elevation_angle_0",
-                        "sw_version", "terrain_correction", "tile_id", "view_angle", "view_angle_0",
-                        "view_angle_1")
 
 
 def as_json_string(str_or_dict):
